+++ ext/geos_c_impl/geometry.c
@@ -1121,6 +1121,43 @@ static VALUE method_geometry_point_on_surface(VALUE self)
   return result;
 }
 
+/**
+  * call-seq:
+  *   some.polygonize -> RGeo::Feature::GeometryCollection
+  *
+  * Polygonizes a set of Geometries which contain linework that
+  * represents the edges of a planar graph.
+  *
+  * All types of Geometry are accepted as input;
+  * the constituent linework is extracted as the edges to be polygonized.
+  * 
+  * The edges must be correctly noded;
+  * that is, they must only meet at their endpoints and not overlap anywhere.
+  *
+  * @see https://libgeos.org/doxygen/geos__c_8h.html#a9d98e448d3b846d591c726d1c0000d25 GEOSPolygonize
+  */
+static VALUE method_geometry_polygonize(VALUE self)
+{
+  VALUE result;
+  RGeo_GeometryData* self_data;
+  const GEOSGeometry* self_geom;
+  GEOSGeometry* geos_polygon_collection;
+
+  result = Qnil;
+  self_data = RGEO_GEOMETRY_DATA_PTR(self);
+  self_geom = self_data->geom;
+  if (self_geom) {
+    geos_polygon_collection = GEOSPolygonize_r(self_data->geos_context, &self_geom, 1);
+
+    if (geos_polygon_collection == NULL) {
+      rb_raise(rb_eGeosError, "GEOS can't polygonize this geometry.");
+    }
+
+    result = rgeo_wrap_geos_geometry(self_data->factory, geos_polygon_collection, Qnil);
+  }
+  return result;
+}
+
 VALUE rgeo_geos_geometries_strict_eql(GEOSContextHandle_t context, const GEOSGeometry* geom1, const GEOSGeometry* geom2)
 {
   switch (GEOSEqualsExact_r(context, geom1, geom2, 0.0)) {
@@ -1190,6 +1227,7 @@ void rgeo_init_geos_geometry()
   rb_define_method(geos_geometry_methods, "invalid_reason_location", method_geometry_invalid_reason_location, 0);
   rb_define_method(geos_geometry_methods, "point_on_surface", method_geometry_point_on_surface, 0);
   rb_define_method(geos_geometry_methods, "make_valid", method_geometry_make_valid, 0);
+  rb_define_method(geos_geometry_methods, "polygonize", method_geometry_polygonize, 0);
 }