+++ test/geos_capi/line_string_test.rb
@@ -22,4 +22,32 @@ class GeosLineStringTest < Minitest::Test # :nodoc:
     interpolated_point = line_string.interpolate_point location
     assert_equal point, interpolated_point
   end
+
+  def test_polygonize_valid_ring
+    input = @factory.parse_wkt("LINESTRING(0 0, 1 1, 1 0, 0 0)")
+    expected = @factory.parse_wkt("GEOMETRYCOLLECTION(POLYGON ((0 0, 1 1, 1 0, 0 0)))")
+
+    assert_equal expected, input.polygonize
+  end
+
+  def test_polygonize_not_closed_ring
+    input = @factory.parse_wkt("LINESTRING(0 0, 1 1, 1 0)")
+    expected = @factory.parse_wkt("GEOMETRYCOLLECTION EMPTY")
+
+    assert_equal expected, input.polygonize
+  end
+
+  def test_polygonize_self_intersection
+    input = @factory.parse_wkt("LINESTRING(0 0, 1 1, 1 0, 0 1, 0 0)")
+    expected = @factory.parse_wkt("GEOMETRYCOLLECTION EMPTY")
+
+    assert_equal expected, input.polygonize
+  end
+
+  def test_polygonize_dangle
+    input = @factory.parse_wkt("LINESTRING(0 0, 2 2, 1 1, 1 0, 0 0)")
+    expected = @factory.parse_wkt("GEOMETRYCOLLECTION EMPTY")
+
+    assert_equal expected, input.polygonize
+  end
 end if RGeo::Geos.capi_supported?