+++ test/geos_capi/multi_point_test.rb
@@ -14,4 +14,11 @@ class GeosMultiPointTest < Minitest::Test # :nodoc:
   def create_factory(opts = {})
     RGeo::Geos.factory(opts)
   end
+
+  def test_polygonize
+    input = @factory.parse_wkt("MULTIPOINT ((1 1))")
+    expected = @factory.parse_wkt("GEOMETRYCOLLECTION EMPTY")
+
+    assert_equal expected, input.polygonize
+  end
 end if RGeo::Geos.capi_supported?